test(CarTest): 전진, 테스트 작성 및 Car 클래스 리팩토링
