--- src/test/java/racingcar/domain/Car.java
@@ -1,18 +1,0 @@
-package racingcar.domain;
-
-public class Car {
-
-    private String name;
-
-    public Car(String name) {
-        this.name = name;
-    }
-
-    public void moveForward(int no) {
-
-    }
-
-    public int getMoveCount() {
-        return 0;
-    }
-}
